+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Procedure when bumping the version number:
 ### Added
 - New messages are now marked as unseen
 - Sub-trees can now be folded
+- Key bindings to move to prev/next sibling
 
 ### Changed
 - Improved editor key bindings